promdump: Implement backoff for batch size
This diff adds a backoff mechanism for batch size.

If Prometheus receives a query that would produce more than a certain number of samples, it will fail the query to protect itself from running out of memory and crashing. The Prometheus API returns HTTP 422 in these cases.

Unfortunately, the Prometheus golang client swallows the HTTP response code, which leaves us with limited options for detecting this condition. The code in this diff uses a RegEx match to scrape the error message but that's probably fine because this is the kind of utility where we can iterate rapidly if a change in the message breaks the scraping.

If the "too many samples" condition occurs, the batch duration will be halved, the existing data will be discarded, and the dump will be restarted from the beginning. If the batch size ever falls below the minimum of 1 second, the utility will throw a fatal error and exit.