promdump: multi-metric support functionally complete
Completed the implementation of flag handling combining custom --metric specifiers and flags for enabling or disabling collection of the Yugabyte metrics exported under master_export, node_export, tserver_export, cql_export (not a typo), and ysql_export.

If only --node_prefix is specified, Yugabyte metrics will be collected using the default export name for each metric (i.e. what is specified in export_type in the PromQL query).

If --metric is specified but --node_prefix is not, only the specified metric will be collected.

If both --node_prefix and --metric are specified, both the Yugabyte metrics and the specified custom metric will be collected.

Individual Yugabyte metrics can be enabled or disabled using the corresponding flags. These flags are enabled by default when node_prefix is specified and disabled by default when it is not. The flag list is generated dynamically based on the collector configuration specified in the var() section of the code. New export collectors utilizing metrics of the same format (i.e. {export_type="...",node_prefix="..."}) can be added with no additional code required.

Code cleanup is required before merging but I want to get this functionally complete prototype pushed out to an offsite repo ASAP.