promdump: fix handling of future end times
While calculating time bounds from the provided flags, there was an issue where a future end time would be truncated to the current time but the corresponding query period would not be properly reset. This led to the utility attempting to run queries with negative durations during the export step. This commit fixes the issue by recalculating the period if the end time is truncated. Added a corresponding log message.

We may lose one sample per metric due to rounding instead of using a ceiling function but the odds of this are low and the loss of a single sample at the end of the data set is normally not of particular consequence.

Refactored the check for existing files to use a func. I'm not sure whether the new implementation is strictly better but at least the logic of the check itself is no longer duplicated.

Added a debug flag and put some of the chattier log messages behind it.