promdump: Add --node_prefix validation and make this flag optional when --instances is specified
Implements https://github.com/yugabyte/yb-tools/issues/93.

Previously, the --node_prefix flag was still required even when specifying a --instances flag. This commit adds logic to relax this requirement when it is safe to do so.

This commit also adds some basic validation for the --node_prefix flag and an override flag --node_prefix_validation to disable the validation if necessary. This validation should reduce the number of issues where customers specify an incorrect node prefix and only the platform metrics are collected.

Inverted the logic of the ambiguously named "isDefault" boolean in the promExport struct and changed its name to "changedFromDefault" to better reflect its purpose. This boolean flags any metrics that have been manually toggled by the user, so that the "collect" boolean isn't overridden by code that contextually flips metrics collectors on or off.

Added a condition to omit the node_prefix PromQL selector if the user has specified --instances instead.

Cleaned up several stale comments.