promdump: integrate with the YBA API to look up Universes by name or UUID
The major change in this commit is YBA API integration. This feature is intended to address the most common issue we see with promdump, where customers do not provide the correct --node_prefix value and the Yugabyte metrics don't get collected. When this happens, there has to be another round-trip with the customer to get the metrics.

Updates go.mod files to rename the module from knyar/prometheus-remote-backfill to yugabyte/prometheus-remote-backfill

Lots of changes to the flags.

* Added the --list_universes flag that will (along with the appropriate hostname and API token values) query the YBA API for the list of Universes, print it to the console, and exit. There's room for some kind of interactive menu here but that's for another day.
* Added the --universe_name and --universe_uuid flags that will instruct promdump to use the YBA API to retrieve the correct node prefix from YBA.
* Marked the --node_prefix flag as deprecated.
* Added the --yba_api_hostname and --yba_api_token flags for specifying the values required to call the YBA API
* Added the --yba_api_timeout, --yba_api_use_tls, and --yba_skip_host_verification flags to control YBA API connection behaviour. Many customers are using self-signed certificates, so the ability to skip host validation is important.

The basic flow for the new functionality is that if the user has provided the --list_universes, --universe_name, --universe_uuid, or --yba_token flags, the YBA API feature will be activated.

We then validate the flags to confirm the following:
* A token has been provided.
* The user has not also provided a node prefix. This was deliberately made a fatal error to encourage a complete move to the new functionality and to avoid user confusion if the provided node prefix is ignored.
* Either a Universe name or Universe UUID has been provided but not both.
* The Universe UUID looks like a UUID.

The utility will print warnings when using insecure settings.

We create a context called ybaCtx that is used to pass the YBA API token to the API calls.

The setupYBAAPI() func is used to configure other client settings such as TLS, certificate verification, client timeouts, etc. and the resulting client handle is passed back to main.

We than call the getCustomerUuid func, passing in the context and client handle. This makes the API call and handles the response, returning the customer UUID. If multiple customer UUIDs are found, promdump exits with a fatal because the utility does not currently support multi-tenant environments.

If the user has asked for the list of Universes, we call the API to retrieve the full list of Universes, then pretty print the list and exit.

If the user has specified a universe by name or UUID, we call the corresponding YBA API function to retrieve a handle to the Universe response object. If the lookup fails (e.g. because there is no Universe that matches the search criterion), promdump will print the list of Universes and exit with a fatal error.

Once we have the correct universe object, we retrieve its node prefix attribute and write it into the nodePrefix variable.

The remainder of the process for retrieving and writing the metrics to disk is unchanged.

The custom metric functionality exposed using --metric and --out is unchanged and remains fully supported because we still need a mechanism for collecting custom metrics.