promdump: add the ability to skip TLS cert verification for Prometheus
Added a flag --skip_prometheus_host_verification to allow promdump to ignore the validity of the host certificate for Prometheus servers where TLS is enabled.

Added --prometheus_api_timeout to control the HTTP timeout for Prometheus API calls since we were in the neighborhood.

The Prometheus API client handle is now created by a func called setupPromAPI that handles initialization of the HTTP client and Prometheus API client, then returns a handle to the new API client.