promdump: implement logging to file
The utility will now record its log messages to a log file on disk (promdump.log by default). This log file will be appended to if it exists.

In addition, the log file will be bundled into the tar bundle at the last possible moment. The log will still be truncated but will include messages written up until that point.