promdump: record the flags provided by the user
This commit adds a func that loops through the list of flags parsed by the flag module and records each flag and value in the log.

The flags are joined using a string builder and logged as a single line for ease of greppage. Fixed typo in log message.

The yba_api_token flag has special handling to avoid recording the token value. This masking uses a generalized mechanism, so sensitive flags added in future can be trivially masked by adding them to the sensitiveFlags slice by name.