promdump: add getRangeTimestamps tests to the test suite
This update to the test suite adds test cases for the following:
* at most two of startTime, endTime, and period can be specified
* endTime should default to now if neither the start time nor end time are specified
* endTime should default to now if only the period is specified
* when both start time and end time are specified, the period is the difference between the two
* when none of start time, end time, or period are specified, the period should be the default period
* when we have the period and one time stamp, the other time stamp should be time stamp +/- the period
* startTime must be in the past
* startTime must be before endTime
* end timestamp should be set to now if it's a time from the future

Implementing tests for behaviour that depends on the current time required some changes to the core promdump. All calls to time.Now() were replaced with a func called now() that returns time.Now().

In the test suite, there is a func called freezeTime() that gets the current time from time.Now() and overrides the now() func so that it always returns this static time. This ensures that tests that rely on now() are deterministic.