promdump: update test suite to be more declarative
Updating the test suite to a more declarative style.

The list of tests to run, their arguments and expected returns are now built using a set of structs (though pointers need to be populated separately), then executed in a slightly cryptic foreach-style "range" loop.

This approach is cleaner but does come at a cost of reduced flexibility in case we have tests that need to be "different" for some reason.