promdump: fix crash in error handling
The call to promRetryWait to calculate the sleep time was moved before the calls that check whether a particular error is fatal, retryable, etc.. This led to the actual error we're trying to test being masked by the error return from promRetryWait, leading to a nil error being passed into promErrIssFatal and causing a crash.

This commit changes the name of the variable holding the error return from promRetryWait in order to prevent the crash.