promdump: custom metrics are broken
The error handling changes made so that the tool aborts if none of the metrics were dumped successfully did not consider custom metrics, so custom metric export would always fail.

This fix moves the skipped, successful, and failed metrics counts to the global scope and increments these counts appropriately during custom metric export.

The summary line was updated to use a calculated total that includes the custom metric if specified.

This commit also makes failure to export a custom metric (if specified) non-fatal.