Remove promremotewrite from the repo
The promremotewrite utility depends on a Prometheus library that in turn depends on the gogoprotobuf library. This library is deprecated, has not been updated in more than two years, and has known security vulnerabilities.

Since this utility is not used at Yugabyte, this made the decision about what to do about it pretty easy.