Implement minimal collection mode
This commit adds the capability to reduce metrics collection to just a critical subset of the metrics. The set of metrics to be collected is derived from the YBA "MINIMAL" metrics collection level.

Add a constant minimalCollectionPromRE that stores the string for the PromQL regular expression used to filter metrics in MINIMAL collection mode.

Add the --collection_level flag which accepts the values "normal" (the default) or "minimal".

Add code to validate the --collection_level flag value.

Add handling for the new --collection_level flag to the PromQL metric builder. This adds a RegEx match expression for minimalCollectionPromRE on the saved_name label when the collection level is set to minimal and we are processing the tserver export.

Update README.md with documentation for the new flag.