Add tserver metric filtering by type
Most of the metrics for the tserver export have a metric_type label. This label - if present - is one of "table", "tablet", "server". This commit adds the ability to filter based on this metric type.

Update the description of the --collection_level flag to indicate that it is mutually exclusive with the new --skip flags that control the metric type filtering.

Add the --skip_table_metrics, --skip_tablet_metrics, --skip_tserver_server_metrics, and --skip_tserver_untyped_metrics flags.

Add validation code to enforce the mutual exclusivity of the --collection_level and --skip flags.

Add code for each skip flag that appends a metric_type label specifier to the PromQL query excluding the metric type to be skipped. Log that a metric type has been skipped.

Update README.md to rewrite the Collection Level section into a more general Tserver Export Filtering section instead. Add the new skip flags and their details to the flag list.