Add debug build that includes symbols
The GitHub actions that build the promdump binary will now produce two versions of the binary, one with debug symbols and one without. These symbols are only required for analyzing crashes, which are rare in promdump.

A SymbolStatus variable has been added to the code that can be overridden at build time. The two versions of the binary are differentiated in the version string as STRIPPED vs NON-STRIPPED.